refactor: reorganise content component
