--- src/interface/newslistprops.ts
@@ -1,7 +1,0 @@
-import NewsApiRequest from './newsapirequest';
-
-interface NewsListProps extends NewsApiRequest {
-  setTotalPages: (p: number) => void;
-}
-
-export default NewsListProps;
